+++ R/variogram.r
@@ -3,11 +3,14 @@
 #' Produces variogram plots for checking spatial trends.
 #'
 #' @param model.obj An `asreml` model object.
-#' @param palette A string specifying the colour scheme to use for plotting. The default value (`"default"`) is equivalent to `"rainbow"`. Colour blind friendly palettes can also be provided via options `"colour blind"` (or `"color blind"`, both equivalent to `"viridis"`), `"magma"`, `"inferno"`, `"plasma"` or `"cividis"`. The `"Spectral"` palette from [scales::brewer_pal()] is also possible.
+#' @param row A row variable.
+#' @param column A column variable.
+#' @param horizontal Logical (default `TRUE`). The direction the plots are arranged. The default `TRUE` places the plots above and below, while `FALSE` will place them side by side.
+#' @param palette A string specifying the colour scheme to use for plotting. The default value (`"default"`) is equivalent to `"rainbow"`. Colour blind friendly palettes can also be provided via options `"colo(u)r blind"` (both equivalent to `"viridis"`), `"magma"`, `"inferno"`, `"plasma"`, `"cividis"`, `"rocket"`, `"mako"` or `"turbo"`. The `"Spectral"` palette from [scales::brewer_pal()] is also possible.
 #'
-#' @return A ggplot2 object.
+#' @return A `ggplot2` object.
 #'
-#' @importFrom interp interp
+#' @importFrom pracma interp2
 #' @importFrom grDevices rainbow
 #' @importFrom lattice wireframe
 #' @importFrom cowplot plot_grid
@@ -29,27 +32,41 @@
 #' }
 #' @export
 
-variogram <- function(model.obj, palette = "default") {
+variogram <- function(model.obj, row = NA, column = NA, horizontal = TRUE, palette = "default") {
 
     if(!(inherits(model.obj, "asreml"))) {
         stop("model.obj must be an asreml model object")
     }
 
-    aa <- vario_df(model.obj)
-    xnam <- names(aa)[2]
-    ynam <- names(aa)[1]
-    fld <- interp::interp(y = aa[,1], x = aa[,2], z = aa$gamma)
-    gdat <- cbind(expand.grid(x = fld$x, y = fld$y), z = as.vector(fld$z))
+    aa <- vario_df(model.obj, row, column)
 
-    a <- ggplot2::ggplot(gdat, ggplot2::aes(x = y, y = x, z = z, fill = z)) +
-        ggplot2::geom_tile(alpha = 0.6) +
+    if(missing(row) | is.na(row) | is.null(row)) {
+        row <- names(aa)[2]
+    }
+    if(missing(column) | is.na(column) | is.null(column)) {
+        column <- names(aa)[1]
+    }
+    row_vals <- unique(aa[,2]) # x
+    col_vals <- unique(aa[,1]) # y
+    z <- matrix(aa$gamma, nrow = length(row_vals), byrow = TRUE)
+
+    interp_rows <- seq(min(row_vals), max(row_vals), length = 40)
+    interp_cols <- seq(min(col_vals), max(col_vals), length = 40)
+    gdat <- expand.grid(x = interp_rows, y = interp_cols)
+
+    pr <- pracma::interp2(x = col_vals, y = row_vals, Z = z, xp = gdat$y, yp = gdat$x)
+    pr <- matrix(pr, nrow = length(interp_rows), byrow = F)
+    gdat <- cbind(gdat, z = as.vector(pr))
+
+    a <- ggplot2::ggplot(gdat, ggplot2::aes(x = y, y = x, z = z)) +
+        ggplot2::geom_tile(alpha = 0.6, ggplot2::aes(fill = z)) +
         ggplot2::coord_equal() +
         ggplot2::geom_contour(color = "white", alpha = 0.5) +
         ggplot2::theme_bw(base_size = 8) +
         ggplot2::scale_y_continuous(expand = c(0, 0), breaks = seq(1, max(gdat$x), 2)) +
         ggplot2::scale_x_continuous(expand = c(0, 0), breaks = seq(1, max(gdat$y), 2)) +
         ggplot2::theme(legend.position = "none", aspect.ratio = 0.3) +
-        ggplot2::labs(y = paste(xnam, "Lag", sep = " "), x = paste(ynam, "Lag", sep = " "))
+        ggplot2::labs(y = paste(row, "Lag", sep = " "), x = paste(column, "Lag", sep = " "))
 
     if(tolower(palette) == "rainbow" | tolower(palette) == "default") {
         a <- a + ggplot2::scale_fill_gradientn(colours = grDevices::rainbow(100))
@@ -58,8 +75,8 @@ variogram <- function(model.obj, palette = "default") {
                                 scales = list(cex = 0.5, arrows = FALSE),
                                 shade = TRUE, colorkey = FALSE,
                                 par.settings = list(axis.line = list(col = 'transparent')),
-                                xlab = list(label = paste(ynam, "Lag", sep = " "), cex = .8, rot = 20),
-                                ylab = list(label = paste(xnam, "Lag", sep = " "), cex = .8, rot = -18),
+                                xlab = list(label = paste(column, "Lag", sep = " "), cex = .8, rot = 20),
+                                ylab = list(label = paste(row, "Lag", sep = " "), cex = .8, rot = -18),
                                 zlab = list(label = NULL, cex.axis = 0.5))
     }
     else if(any(grepl("(colou?r([[:punct:]]|[[:space:]]?)blind)|cb|viridis", palette, ignore.case = T))) {
@@ -70,12 +87,12 @@ variogram <- function(model.obj, palette = "default") {
                                 scales = list(cex = 0.5, arrows = FALSE),
                                 drape = TRUE, colorkey = FALSE,
                                 par.settings = list(axis.line = list(col = 'transparent')),
-                                xlab = list(label = paste(ynam, "Lag", sep = " "), cex = .8, rot = 20),
-                                ylab = list(label = paste(xnam, "Lag", sep = " "), cex = .8, rot = -18),
+                                xlab = list(label = paste(column, "Lag", sep = " "), cex = .8, rot = 20),
+                                ylab = list(label = paste(row, "Lag", sep = " "), cex = .8, rot = -18),
                                 zlab = list(label = NULL, cex.axis = 0.5),
                                 col.regions = scales::viridis_pal(option = "viridis")(100))
     }
-    else if(tolower(palette) %in% c("magma", "inferno", "cividis", "plasma")) {
+    else if(tolower(trimws(palette)) %in% c("magma", "inferno", "cividis", "plasma", "rocket", "mako", "turbo")) {
         a <- a + ggplot2::scale_fill_gradientn(colours = scales::viridis_pal(option = palette)(50))
 
         # Create the lattice plot
@@ -83,13 +100,13 @@ variogram <- function(model.obj, palette = "default") {
                                 scales = list(cex = 0.5, arrows = FALSE),
                                 drape = TRUE, colorkey = FALSE,
                                 par.settings = list(axis.line = list(col = 'transparent')),
-                                xlab = list(label = paste(ynam, "Lag", sep = " "), cex = .8, rot = 20),
-                                ylab = list(label = paste(xnam, "Lag", sep = " "), cex = .8, rot = -18),
+                                xlab = list(label = paste(column, "Lag", sep = " "), cex = .8, rot = 20),
+                                ylab = list(label = paste(row, "Lag", sep = " "), cex = .8, rot = -18),
                                 zlab = list(label = NULL, cex.axis = 0.5),
                                 col.regions = scales::viridis_pal(option = palette)(100))
     }
 
-    else if(palette %in% c("Spectral")) {
+    else if(tolower(trimws(palette)) %in% c("spectral")) {
         a <- a + ggplot2::scale_fill_gradientn(colours = scales::brewer_pal(palette = palette)(11))
 
         # Create the lattice plot
@@ -97,8 +114,8 @@ variogram <- function(model.obj, palette = "default") {
                                 scales = list(cex = 0.5, arrows = FALSE),
                                 drape = TRUE, colorkey = FALSE,
                                 par.settings = list(axis.line = list(col = 'transparent')),
-                                xlab = list(label = paste(ynam, "Lag", sep = " "), cex = .8, rot = 20),
-                                ylab = list(label = paste(xnam, "Lag", sep = " "), cex = .8, rot = -18),
+                                xlab = list(label = paste(column, "Lag", sep = " "), cex = .8, rot = 20),
+                                ylab = list(label = paste(row, "Lag", sep = " "), cex = .8, rot = -18),
                                 zlab = list(label = NULL, cex.axis = 0.5),
                                 col.regions = scales::brewer_pal(palette = palette)(11))
     }
@@ -123,19 +140,40 @@ variogram <- function(model.obj, palette = "default") {
 #'
 #' @param model.obj An asreml model
 #'
-#' @return A data frame with the variogram for a model. The data frame contains the spatial coordinates (typically row and column), the $gamma$ for that position and the number of points with the separation.
+#' @return A data frame with the variogram for a model. The data frame contains the spatial coordinates (typically row and column), the `gamma` for that position and the number of points with the separation.
 #' @keywords internal
-vario_df <- function(model.obj) {
+#'
+#'
+#' @examples
+#' \dontrun{
+#' library(asreml)
+#' oats <- asreml::oats
+#' oats <- oats[order(oats$Row, oats$Column),]
+#' model.asr <- asreml(yield ~ Nitrogen + Variety + Nitrogen:Variety,
+#'                     random = ~ Blocks + Blocks:Wplots,
+#'                     residual = ~ ar1(Row):ar1(Column),
+#'                     data = oats)
+#' vario_df(model.asr)
+#' }
+#'
+vario_df <- function(model.obj, Row = NA, Column = NA) {
     # The 'z' value for the variogram is the residuals
     # Need to be able to pull out the x/y from the model object
 
     dims <- unlist(strsplit(names(model.obj$R.param[1]), ":"))
-    Row <- as.numeric(model.obj$mf[[dims[1]]])
-    Column <- as.numeric(model.obj$mf[[dims[2]]])
-    Resid <- residuals(model.obj)
+
+    if(missing(Row) | is.na(Row) | is.null(Row)) {
+        Row <- as.numeric(model.obj$mf[[dims[1]]])
+    }
+    if(missing(Column) | is.na(Column) | is.null(Column)) {
+        Column <- as.numeric(model.obj$mf[[dims[2]]])
+    }
 
     nrows <- max(Row)
     ncols <- max(Column)
+
+    Resid <- residuals(model.obj)#[order(Column, Row)], nrow = nrows)
+    # Resid <- matrix(residuals(model.obj)[order(Column, Row)], nrow = nrows)
 
     vario <- expand.grid(Row = 0:(nrows-1), Column = 0:(ncols-1))
 
@@ -149,7 +187,7 @@ vario_df <- function(model.obj) {
 
         gamma <- 0
         np <- 0
-        for (val_index in 1:nrows) {
+        for (val_index in 1:nrow(vario)) {
             # val <- vals[val_index, ]
 
             # Deliberate double-counting so that offset handling is easy
@@ -160,8 +198,10 @@ vario_df <- function(model.obj) {
                 col <- Column[val_index] + offset[2]
 
                 if (0 < row && row <= nrows && 0 < col && col <= ncols) {
-                    other <- which(Row == row & Column == col)
-                    gamma <- gamma + (Resid[val_index]-Resid[other])^2
+                    other <- ifelse(!is.na(Resid[Row == row & Column == col]),
+                                    Resid[Row == row & Column == col],
+                                    0)
+                    gamma <- gamma + (Resid[val_index]-other)^2
                     np <- np + 1
                 }
             }