+++ NEWS.md
@@ -1,3 +1,21 @@
+# biometryassist 1.1.1
+
+## Minor changes
+
+- `multiple_comparisons()` now accepts power transformations and automatically back-transforms. It gains a new argument `power` to provide the transformation power applied in the model to undo. This enables more general Box-Cox transformations. (#36)
+- `multiple_comparisons()` no longer produces an error when the `trans` argument is supplied and `offset` is not. It now produces a warning and sets `offset` to 0 when not provided. (#37)
+- Added an option to turn off the start up message and version check. Add `options(biometryassist.check = FALSE)` to your .Rprofile file to disable. Partially fixes #6.
+- Enabled new colour-blind friendly palettes (#39)
+
+## Bug fixes
+
+- Updated the required version of rlang (>=1.0.0)
+- Fixed a bug that didn't allow labels and the x axis to be rotated independently for `autoplot.mct()` (#35)
+- Fixed an issue where treatment columns were not determined correctly in `multiple_comparisons()` (#33)
+- Fixed a problem with variograms displaying grey values where `NA`s have been produced. (#24 and #25) 
+- Fixed some warnings that would be introduced with the forthcoming version of ggplot2 (#42)
+
+
 # biometryassist 1.1.0
 
 ## Major changes